+++ src/AppBundle/Form/Type/UserType.php
@@ -1,26 +1,21 @@
 <?php
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type as FormType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class GameType extends AbstractType {
+class UserType extends AbstractType {
 
     /**
      * {@inheritdoc}
      */
     public function buildForm(FormBuilderInterface $builder, array $options) {
-        $builder->add('name', FormType\TextType::class)
-                ->add('nbPlayers', FormType\ChoiceType::class, array('choices' => array(
-                        '1' => '1',
-                        '2' => '2',
-                        '3' => '3',
-                        '4' => '4')))
-                ->add('create', FormType\SubmitType::class, array('label' => 'create'))
-                ->getForm();
+        $builder->add('login')
+                ->add('password', FormType\PasswordType::class)
+                ->add('save', FormType\SubmitType::class, array('label' => 'ok'));
     }
 
     /**
@@ -28,7 +23,7 @@ class GameType extends AbstractType {
      */
     public function configureOptions(OptionsResolver $resolver) {
         $resolver->setDefaults(array(
-            'data_class' => 'AppBundle\Entity\Game'
+            'data_class' => 'AppBundle\Entity\User'
         ));
     }
 
@@ -36,7 +31,7 @@ class GameType extends AbstractType {
      * {@inheritdoc}
      */
     public function getBlockPrefix() {
-        return 'appbundle_game';
+        return 'appbundle_user';
     }
 
 }
